# Dmatrix optimization2