Observed #5530 slowdown in DMatrix creation was fixed.

Performance improvements:

DMatrix create time,s | Master | Optimized | Speed up | Optimized without specialization | Speed up
-- | -- | -- | -- | -- | --
airline-ohe contiguous numpy | 6.6 | 1.8 | 3.67 | 1.99 | 3.32
msrank10k | 3.9 | 1.16 | 3.36 | 1.61 | 2.42
mortgage1Q | 4.06 | 1.1 | 3.69 | 2.93 | 1.39